+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2026-07-17T13:55:55.639Z\n"
-"PO-Revision-Date: 2026-07-17T13:55:55.639Z\n"
+"POT-Creation-Date: 2026-07-17T19:19:08.730Z\n"
+"PO-Revision-Date: 2026-07-17T19:19:08.731Z\n"
 
 msgid "2020"
 msgstr "2020"
@@ -155,29 +155,6 @@ msgstr "Operator"
 msgid "Date"
 msgstr "Date"
 
-msgid "Selected"
-msgstr "Selected"
-
-msgid "Not selected"
-msgstr "Not selected"
-
-msgid "All"
-msgstr "All"
-
-msgid "{{count}} selected"
-msgid_plural "{{count}} selected"
-msgstr[0] "{{count}} selected"
-msgstr[1] "{{count}} selected"
-
-msgid "No features match your filters"
-msgstr "No features match your filters"
-
-msgid "Clear filters"
-msgstr "Clear filters"
-
-msgid "No results found"
-msgstr "No results found"
-
 msgid "Select all visible rows"
 msgstr "Select all visible rows"
 
@@ -241,6 +218,20 @@ msgstr "No matches"
 msgid "No value"
 msgstr "No value"
 
+msgid "Selected"
+msgstr "Selected"
+
+msgid "Not selected"
+msgstr "Not selected"
+
+msgid "All"
+msgstr "All"
+
+msgid "{{count}} selected"
+msgid_plural "{{count}} selected"
+msgstr[0] "{{count}} selected"
+msgstr[1] "{{count}} selected"
+
 msgid "Drill up one level"
 msgstr "Drill up one level"
 
@@ -262,6 +253,15 @@ msgstr "Zoom to selected features"
 msgid "Zoom to filtered features"
 msgstr "Zoom to filtered features"
 
+msgid "No features match your filters"
+msgstr "No features match your filters"
+
+msgid "Clear filters"
+msgstr "Clear filters"
+
+msgid "No results found"
+msgstr "No results found"
+
 msgid "Close"
 msgstr "Close"
 
@@ -271,15 +271,6 @@ msgstr "Restore"
 msgid "Collapse"
 msgstr "Collapse"
 
-msgid "Drag to reorder"
-msgstr "Drag to reorder"
-
-msgid "Unpin column"
-msgstr "Unpin column"
-
-msgid "Pin column to the left"
-msgstr "Pin column to the left"
-
 msgid "Configure columns"
 msgstr "Configure columns"
 
@@ -288,6 +279,15 @@ msgstr "Select all columns"
 
 msgid "Reset to defaults"
 msgstr "Reset to defaults"
+
+msgid "Drag to reorder"
+msgstr "Drag to reorder"
+
+msgid "Unpin column"
+msgstr "Unpin column"
+
+msgid "Pin column to the left"
+msgstr "Pin column to the left"
 
 msgid "Search across all visible columns"
 msgstr "Search across all visible columns"
