+++ src/components/datatable/DataTable.jsx
@@ -1,16 +1,11 @@
 import i18n from '@dhis2/d2-i18n'
 import {
-    DataTable,
     DataTableRow,
     DataTableCell,
     DataTableColumnHeader,
-    DataTableHead,
-    DataTableBody,
     ComponentCover,
     CenteredContent,
     CircularLoader,
-    Popper,
-    Portal,
     IconSync16,
 } from '@dhis2/ui'
 import cx from 'classnames'
@@ -34,279 +29,33 @@ import {
 import {
     SENTINEL_SELECTED_ROW,
     SORT_ASCENDING,
-    SORT_DESCENDING,
 } from '../../constants/dataTable.js'
-import {
-    SELECTION_FILTER_SELECTED,
-    SELECTION_FILTER_NOT_SELECTED,
-} from '../../constants/selection.js'
 import { isDarkColor } from '../../util/colors.js'
+import {
+    getNextSorting,
+    getRowClickAction,
+    getRowId,
+    isFilterable,
+    shouldClearFeatureHighlight,
+} from '../../util/dataTable.js'
 import { formatWithSeparator } from '../../util/numbers.js'
 import {
+    getPinnedCellProps,
+    getPinnedCount,
     getPinnedLeftOffsets,
     getVisibleHeaders,
 } from '../../util/tableColumns.js'
 import { useCachedData } from '../cachedDataProvider/CachedDataProvider.jsx'
-import Checkbox from '../core/Checkbox.jsx'
 import { SortIcon } from '../core/icons.jsx'
-import {
-    FilterDropdownPopover,
-    getDropdownPlacement,
-} from './FilterDropdownPopover.jsx'
 import FilterInput from './FilterInput.jsx'
+import SelectionFilterButton from './SelectionFilterButton.jsx'
 import styles from './styles/DataTable.module.css'
 import TableContextMenu from './TableContextMenu.jsx'
+import TableComponents from './TableVirtuosoComponents.jsx'
+import TopTooltip from './TopTooltip.jsx'
 import { useColumnWidths } from './useColumnWidths.js'
 import { useRowSelection } from './useRowSelection.js'
 import { useTableData } from './useTableData.js'
-
-const SELECTION_FILTER_OPTIONS = [
-    { value: SELECTION_FILTER_SELECTED, label: i18n.t('Selected') },
-    { value: SELECTION_FILTER_NOT_SELECTED, label: i18n.t('Not selected') },
-]
-
-export const isFilterable = (dataKey, type) => !!type
-
-const SelectionFilterButton = ({ value, onChange }) => {
-    const anchorRef = useRef(null)
-    const [isOpen, setIsOpen] = useState(false)
-
-    const toggleValue = (optionValue) => {
-        const next = value.includes(optionValue)
-            ? value.filter((v) => v !== optionValue)
-            : [...value, optionValue]
-        onChange(next)
-    }
-
-    const buttonLabel =
-        value.length === 0
-            ? i18n.t('All')
-            : i18n.t('{{count}} selected', { count: value.length })
-
-    const anchorRect = anchorRef.current?.getBoundingClientRect()
-    const { dropdownPlacement } = getDropdownPlacement(anchorRect)
-
-    return (
-        <>
-            <button
-                type="button"
-                ref={anchorRef}
-                className={styles.selectionFilterButton}
-                data-test="data-table-selection-filter-button"
-                onClick={() => setIsOpen((o) => !o)}
-            >
-                {buttonLabel}
-            </button>
-            {isOpen && (
-                <FilterDropdownPopover
-                    reference={anchorRef}
-                    placement={dropdownPlacement}
-                    onClickOutside={() => setIsOpen(false)}
-                >
-                    <div className={styles.selectionFilterPopover}>
-                        {SELECTION_FILTER_OPTIONS.map((option) => (
-                            <Checkbox
-                                key={option.value}
-                                label={option.label}
-                                checked={value.includes(option.value)}
-                                onChange={() => toggleValue(option.value)}
-                                style={{ margin: '4px 0' }}
-                            />
-                        ))}
-                    </div>
-                </FilterDropdownPopover>
-            )}
-        </>
-    )
-}
-
-SelectionFilterButton.propTypes = {
-    value: PropTypes.arrayOf(PropTypes.string).isRequired,
-    onChange: PropTypes.func.isRequired,
-}
-
-const topTooltipModifiers = [{ name: 'offset', options: { offset: [0, 4] } }]
-
-const TopTooltip = ({ content, children }) => {
-    const [open, setOpen] = useState(false)
-    const referenceRef = useRef(null)
-    const openTimerRef = useRef(null)
-    const closeTimerRef = useRef(null)
-
-    const onOpen = () => {
-        clearTimeout(closeTimerRef.current)
-        openTimerRef.current = setTimeout(() => setOpen(true), 200)
-    }
-
-    const onClose = () => {
-        clearTimeout(openTimerRef.current)
-        closeTimerRef.current = setTimeout(() => setOpen(false), 200)
-    }
-
-    useEffect(
-        () => () => {
-            clearTimeout(openTimerRef.current)
-            clearTimeout(closeTimerRef.current)
-        },
-        []
-    )
-
-    return (
-        <span
-            ref={referenceRef}
-            onMouseOver={onOpen}
-            onMouseOut={onClose}
-            onFocus={onOpen}
-            onBlur={onClose}
-        >
-            {children}
-            {open && (
-                <Portal>
-                    <Popper
-                        placement="top"
-                        reference={referenceRef}
-                        modifiers={topTooltipModifiers}
-                    >
-                        <div className={styles.topTooltipContent}>
-                            {content}
-                        </div>
-                    </Popper>
-                </Portal>
-            )}
-        </span>
-    )
-}
-
-TopTooltip.propTypes = {
-    children: PropTypes.node.isRequired,
-    content: PropTypes.node.isRequired,
-}
-
-export const shouldClearFeatureHighlight = (event) =>
-    event.relatedTarget?.tagName !== 'TD'
-
-export const getNextSorting = (name, { sortField, sortDirection }) => {
-    if (name !== sortField) {
-        return { sortField: name, sortDirection: SORT_ASCENDING }
-    }
-    if (sortDirection === SORT_ASCENDING) {
-        return { sortField: name, sortDirection: SORT_DESCENDING }
-    }
-    return { sortField: null, sortDirection: SORT_ASCENDING }
-}
-
-const getRowId = (row) =>
-    row.find((r) => r.dataKey === 'id')?.value || row[0]?.itemId
-
-export const getRowClickAction = (
-    event,
-    { id, rowIndex, rows, lastClickedRowIndex }
-) => {
-    if (event.shiftKey) {
-        if (lastClickedRowIndex === null) {
-            return { type: 'toggle', id }
-        }
-        const [start, end] = [lastClickedRowIndex, rowIndex].sort(
-            (a, b) => a - b
-        )
-        const ids = rows
-            .slice(start, end + 1)
-            .map(getRowId)
-            .filter(Boolean)
-        return { type: 'range', ids }
-    }
-
-    if (event.ctrlKey || event.metaKey) {
-        return { type: 'toggle', id }
-    }
-
-    return null
-}
-
-const DataTableWithVirtuosoContext = ({ context, ...props }) => (
-    <DataTable
-        {...props}
-        layout={context.layout}
-        className={styles.dataTable}
-    />
-)
-
-DataTableWithVirtuosoContext.propTypes = {
-    context: PropTypes.shape({
-        layout: PropTypes.string,
-    }),
-}
-
-const DataTableRowWithVirtuosoContext = ({ context, item, ...props }) => (
-    <DataTableRow
-        onMouseEnter={() => context.onMouseEnter(item)}
-        onMouseLeave={context.onMouseLeave}
-        onContextMenu={(e) => context.onContextMenu(e, item)}
-        onClick={(e) => context.onRowClick(item, e)}
-        onDoubleClick={() => context.onRowDoubleClick(item)}
-        {...props}
-    />
-)
-
-DataTableRowWithVirtuosoContext.propTypes = {
-    context: PropTypes.shape({
-        onContextMenu: PropTypes.func,
-        onMouseEnter: PropTypes.func,
-        onMouseLeave: PropTypes.func,
-        onRowClick: PropTypes.func,
-        onRowDoubleClick: PropTypes.func,
-    }),
-    item: PropTypes.arrayOf(
-        PropTypes.shape({
-            dataKey: PropTypes.string,
-            itemId: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-            value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-        })
-    ),
-}
-
-const EmptyPlaceholder = ({ context }) => (
-    <tbody>
-        <tr>
-            <td colSpan={99999}>
-                <div className={styles.noResults}>
-                    {context.totalCount > 0 ? (
-                        <>
-                            {i18n.t('No features match your filters')}
-                            {context.hasActiveFilters && (
-                                <button
-                                    type="button"
-                                    className={styles.clearFiltersLink}
-                                    onClick={context.onClearFilters}
-                                >
-                                    {i18n.t('Clear filters')}
-                                </button>
-                            )}
-                        </>
-                    ) : (
-                        i18n.t('No results found')
-                    )}
-                </div>
-            </td>
-        </tr>
-    </tbody>
-)
-
-EmptyPlaceholder.propTypes = {
-    context: PropTypes.shape({
-        hasActiveFilters: PropTypes.bool,
-        totalCount: PropTypes.number,
-        onClearFilters: PropTypes.func,
-    }),
-}
-
-const TableComponents = {
-    Table: DataTableWithVirtuosoContext,
-    TableBody: DataTableBody,
-    TableHead: DataTableHead,
-    TableRow: DataTableRowWithVirtuosoContext,
-    EmptyPlaceholder,
-}
 
 const Table = ({
     availableWidth,
@@ -451,23 +200,10 @@ const Table = ({
         error,
     })
 
-    // Only the leading columns of visibleHeaders can ever be pinned -
-    // getVisibleHeaders already moves pinned columns to the front - so the
-    // pinned section's size is just how many headers match pinnedKeys
-    // before the first one that doesn't.
-    const pinnedColumnCount = useMemo(() => {
-        if (!pinnedKeys.length || !visibleHeaders) {
-            return 0
-        }
-        let count = 0
-        for (const header of visibleHeaders) {
-            if (!pinnedKeys.includes(header.dataKey)) {
-                break
-            }
-            count++
-        }
-        return count
-    }, [visibleHeaders, pinnedKeys])
+    const pinnedColumnCount = useMemo(
+        () => getPinnedCount(visibleHeaders, pinnedKeys),
+        [visibleHeaders, pinnedKeys]
+    )
 
     const pinnedLeftOffsets = useMemo(
         () => getPinnedLeftOffsets(visibleHeaders, pinnedKeys, columnWidths),
@@ -475,26 +211,7 @@ const Table = ({
     )
     const pinnedOffsetsReady = Object.keys(pinnedLeftOffsets).length > 0
 
-    // The checkbox column only becomes sticky when something else is
-    // actually pinned (and its offset is ready) - otherwise it stays a
-    // plain (non-`fixed`) cell, since @dhis2/ui renders `fixed` cells as
-    // `<th>` rather than `<td>`, which would needlessly change the DOM
-    // shape for the common, nothing-pinned case, and briefly during column
-    // widths being (re)measured after a config change.
     const isCheckboxColumnPinned = pinnedColumnCount > 0 && pinnedOffsetsReady
-
-    // @dhis2/ui requires `width` whenever `fixed` is passed - unpinned
-    // cells keep their existing (unset) width behavior.
-    const getPinnedCellProps = (dataKey, index) => {
-        const leftOffset = pinnedLeftOffsets[dataKey]
-        const isPinned = index < pinnedColumnCount && leftOffset !== undefined
-        return {
-            fixed: isPinned,
-            left: isPinned ? `${leftOffset}px` : undefined,
-            width: isPinned ? `${columnWidths[index] ?? 0}px` : undefined,
-            isLastPinned: index === pinnedColumnCount - 1,
-        }
-    }
 
     useEffect(() => {
         onCountChange?.(totalCount, filteredCount)
@@ -711,7 +428,11 @@ const Table = ({
                         {visibleHeaders.map(
                             ({ name, dataKey, type, optionSet }, index) => {
                                 const { fixed, left, isLastPinned } =
-                                    getPinnedCellProps(dataKey, index)
+                                    getPinnedCellProps(dataKey, index, {
+                                        pinnedLeftOffsets,
+                                        pinnedColumnCount,
+                                        columnWidths,
+                                    })
                                 return (
                                     <DataTableColumnHeader
                                         className={cx(styles.columnHeader, {
@@ -833,7 +554,11 @@ const Table = ({
                                 }
                                 const { value, align } = cell
                                 const { fixed, left, width, isLastPinned } =
-                                    getPinnedCellProps(dataKey, index)
+                                    getPinnedCellProps(dataKey, index, {
+                                        pinnedLeftOffsets,
+                                        pinnedColumnCount,
+                                        columnWidths,
+                                    })
                                 return (
                                     <DataTableCell
                                         key={`dtcell-${dataKey}`}
