+++ src/components/datatable/styles/SelectionFilterButton.module.css
@@ -1,0 +1,32 @@
+.selectionFilterButton {
+    width: 100%;
+    height: 24px;
+    font-size: 11px;
+    padding: 4px 6px;
+    border: 1px solid var(--colors-grey500);
+    border-radius: 3px;
+    box-shadow: inset 0 0 1px 0 rgba(48, 54, 60, 0.1);
+    background: var(--colors-white);
+    cursor: pointer;
+    text-align: left;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+
+.selectionFilterPopover {
+    padding: var(--spacers-dp8);
+    min-width: 140px;
+    background-color: var(--colors-white);
+    border-radius: 4px;
+    box-shadow: var(--elevations-popover);
+}
+
+/* !important beats @dhis2/ui's own Checkbox label font-size. */
+.selectionFilterPopover :global(label) {
+    font-size: 11px !important;
+}
+
+.denseCheckbox {
+    margin: 4px 0;
+}
