+++ src/util/dataTable.js
@@ -1,0 +1,44 @@
+import { SORT_ASCENDING, SORT_DESCENDING } from '../constants/dataTable.js'
+
+export const isFilterable = (dataKey, type) => !!type
+
+export const shouldClearFeatureHighlight = (event) =>
+    event.relatedTarget?.tagName !== 'TD'
+
+export const getNextSorting = (name, { sortField, sortDirection }) => {
+    if (name !== sortField) {
+        return { sortField: name, sortDirection: SORT_ASCENDING }
+    }
+    if (sortDirection === SORT_ASCENDING) {
+        return { sortField: name, sortDirection: SORT_DESCENDING }
+    }
+    return { sortField: null, sortDirection: SORT_ASCENDING }
+}
+
+export const getRowId = (row) =>
+    row.find((r) => r.dataKey === 'id')?.value || row[0]?.itemId
+
+export const getRowClickAction = (
+    event,
+    { id, rowIndex, rows, lastClickedRowIndex }
+) => {
+    if (event.shiftKey) {
+        if (lastClickedRowIndex === null) {
+            return { type: 'toggle', id }
+        }
+        const [start, end] = [lastClickedRowIndex, rowIndex].sort(
+            (a, b) => a - b
+        )
+        const ids = rows
+            .slice(start, end + 1)
+            .map(getRowId)
+            .filter(Boolean)
+        return { type: 'range', ids }
+    }
+
+    if (event.ctrlKey || event.metaKey) {
+        return { type: 'toggle', id }
+    }
+
+    return null
+}
